requirements.txt
langchain
openai
google-search-results
huggingface_hub
wikipedia
fastapi
uvicorn
python-dotenv
